Fix duplicate dismissToast export
